+++ src/main/java/kitchenpos/menu/dto/CreateMenuRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.menu.dto;
 
 import kitchenpos.menu.domain.MenuProduct;
