+++ src/main/java/kitchenpos/menugroup/application/MenuGroupService.java
@@ -2,7 +2,7 @@ package kitchenpos.menugroup.application;
 
 import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.menugroup.repository.MenuGroupRepository;
-import kitchenpos.dto.CreateMenuGroupRequest;
+import kitchenpos.menugroup.dto.CreateMenuGroupRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
