+++ src/main/java/kitchenpos/menugroup/ui/MenuGroupRestController.java
@@ -2,7 +2,7 @@ package kitchenpos.menugroup.ui;
 
 import kitchenpos.menugroup.application.MenuGroupService;
 import kitchenpos.menugroup.domain.MenuGroup;
-import kitchenpos.dto.CreateMenuGroupRequest;
+import kitchenpos.menugroup.dto.CreateMenuGroupRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
