+++ src/main/java/kitchenpos/order/dto/CreateOrderLineItemRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.order.dto;
 
 import kitchenpos.order.domain.OrderLineItem;
