+++ src/main/java/kitchenpos/table/application/TableService.java
@@ -3,9 +3,9 @@ package kitchenpos.table.application;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableValidator;
 import kitchenpos.table.repository.OrderTableRepository;
-import kitchenpos.dto.ChangeNumberOfGuestsRequest;
-import kitchenpos.dto.ChangeOrderTableEmptyRequest;
-import kitchenpos.dto.CreateOrderTableRequest;
+import kitchenpos.table.dto.ChangeNumberOfGuestsRequest;
+import kitchenpos.table.dto.ChangeOrderTableEmptyRequest;
+import kitchenpos.table.dto.CreateOrderTableRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
