+++ src/main/java/kitchenpos/table/dto/ChangeNumberOfGuestsRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.table.dto;
 
 public class ChangeNumberOfGuestsRequest {
