+++ src/main/java/kitchenpos/table/dto/ChangeOrderTableEmptyRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.table.dto;
 
 public class ChangeOrderTableEmptyRequest {
