+++ src/main/java/kitchenpos/table/dto/CreateOrderTableRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.table.dto;
 
 import kitchenpos.table.domain.OrderTable;
