+++ src/main/java/kitchenpos/tablegroup/application/TableGroupService.java
@@ -1,7 +1,7 @@
 package kitchenpos.tablegroup.application;
 
 import kitchenpos.tablegroup.domain.TableGroup;
-import kitchenpos.dto.CreateTableGroupRequest;
+import kitchenpos.tablegroup.dto.CreateTableGroupRequest;
 import kitchenpos.tablegroup.repository.TableGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
