+++ src/main/java/kitchenpos/tablegroup/dto/CreateTableGroupRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.tablegroup.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
