+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -7,8 +7,8 @@ import kitchenpos.product.domain.Product;
 import kitchenpos.menugroup.repository.MenuGroupRepository;
 import kitchenpos.menu.repository.MenuRepository;
 import kitchenpos.product.repository.ProductRepository;
-import kitchenpos.dto.CreateMenuProductRequest;
-import kitchenpos.dto.CreateMenuRequest;
+import kitchenpos.menu.dto.CreateMenuProductRequest;
+import kitchenpos.menu.dto.CreateMenuRequest;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
