+++ src/test/java/kitchenpos/table/application/TableServiceTest.java
@@ -16,9 +16,9 @@ import kitchenpos.order.repository.OrderRepository;
 import kitchenpos.table.repository.OrderTableRepository;
 import kitchenpos.product.repository.ProductRepository;
 import kitchenpos.tablegroup.repository.TableGroupRepository;
-import kitchenpos.dto.ChangeNumberOfGuestsRequest;
-import kitchenpos.dto.ChangeOrderTableEmptyRequest;
-import kitchenpos.dto.CreateOrderTableRequest;
+import kitchenpos.table.dto.ChangeNumberOfGuestsRequest;
+import kitchenpos.table.dto.ChangeOrderTableEmptyRequest;
+import kitchenpos.table.dto.CreateOrderTableRequest;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
