+++ src/test/java/kitchenpos/tablegroup/application/TableGroupServiceTest.java
@@ -15,8 +15,8 @@ import kitchenpos.order.repository.OrderRepository;
 import kitchenpos.table.repository.OrderTableRepository;
 import kitchenpos.product.repository.ProductRepository;
 import kitchenpos.tablegroup.repository.TableGroupRepository;
-import kitchenpos.dto.CreateTableGroupOrderTableRequest;
-import kitchenpos.dto.CreateTableGroupRequest;
+import kitchenpos.tablegroup.dto.CreateTableGroupOrderTableRequest;
+import kitchenpos.tablegroup.dto.CreateTableGroupRequest;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
